+++ src/monitors/index.js
@@ -1,1 +1,5 @@
-module.exports = [require('./waitingpods'), require('./longnotready')];
+module.exports = [
+	require('./waitingpods'),
+	require('./longnotready'),
+	require('./metrics'),
+];